% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaStump-package.R
\docType{package}
\name{adaStump-package}
\alias{adaStump}
\alias{adaStump-package}
\title{c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_title(\"#1\")}",
"adaStump")\Sexpr{tools:::Rd_package_title("adaStump")}}
\description{
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_description(\"#1\")}",
"adaStump")\Sexpr{tools:::Rd_package_description("adaStump")}
}
\details{
The DESCRIPTION file:
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_DESCRIPTION(\"#1\")}",
"adaStump")\Sexpr{tools:::Rd_package_DESCRIPTION("adaStump")}
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_indices(\"#1\")}",
"adaStump")\Sexpr{tools:::Rd_package_indices("adaStump")} ~~ An overview of
how to use the package, including the most important functions ~~
}
\author{
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_author(\"#1\")}",
"adaStump")\Sexpr{tools:::Rd_package_author("adaStump")}

Maintainer:
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_maintainer(\"#1\")}",
"adaStump")\Sexpr{tools:::Rd_package_maintainer("adaStump")}
}

