% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaStump.R
\name{adaStump}
\alias{adaStump}
\title{Fit an ada boosting model with decision stumps as base learners}
\usage{
adaStump(formula, data, ...)
}
\arguments{
\item{formula}{formula object describing the model to be fit, as in
\code{ada} function from "ada" package.}

\item{data}{data.frame object to train the model. The variables named in
formula must be present in this object}

\item{...}{Further arguments to be passed to \emph{ada} function from "ada" package}
}
\value{
An object of class \emph{adaStump} containing:
\item{model }{a data.frame describing the stumps generated by \code{ada}.
The object is normally considerably smaller than the equivalent \code{ada}
object} 
\item{type }{Type of execution performed. It is directly inherited
from the \code{adaStump} call}
}
\description{
This function calls \code{ada} with a default \code{control} argument set to
\code{rpart.control(maxdepth=1,cp=-1,minsplit=0,xval=0)}
}
\examples{
#Load Iris
data(iris)

#Create Variable is Iris as numerical
iris$isSetosa <- as.numeric(iris$Species == "setosa")

#Split sample in 70 train - 30 test
train.ind <- sample(nrow(iris), nrow(iris) * 0.7)

#Train model. For obvious reasons, Species variable is not included in the fit
fit <- adaStump(isSetosa ~ Sepal.Length + Sepal.Width + Petal.Length + Petal.Width, iris[train.ind,], 
                type = "discrete", iter = 10, nu = 0.05, bag.frac = 0.6)
#Prediction
predict(fit,iris[-train.ind,])

}

