% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStumpFrame.R
\name{createStumpFrame}
\alias{createStumpFrame}
\title{Internal - Transfrom ada object into data frame}
\usage{
createStumpFrame(ada.model)
}
\arguments{
\item{ada.model}{model of class "ada" to be pruned}
}
\value{
data.frame containing the stumps information
}
\description{
Internal - Transfrom ada object into data frame
}
\keyword{internal}

