% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlphaProbability.R
\name{getAlphaProbability}
\alias{getAlphaProbability}
\title{Internal - Transform tree value and multiply with alpha according its type}
\usage{
getAlphaProbability(tree.value, alpha, type)
}
\description{
Internal - Transform tree value and multiply with alpha according its type
}
\keyword{internal}

