% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoefficientLine.R
\name{getCoefficientLine}
\alias{getCoefficientLine}
\title{Internal - Get individual tree output from stump.frame object}
\usage{
getCoefficientLine(var, condition, value, probT, probF, alpha = NULL, newdata,
  default, type)
}
\description{
Internal - Get individual tree output from stump.frame object
}
\keyword{internal}

