% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.prunedadaStump.R
\name{predict.prunedadaStump}
\alias{predict.prunedadaStump}
\title{Predict from a \code{prunedadaStump} object}
\usage{
\method{predict}{prunedadaStump}(stump.model, newdata)
}
\arguments{
\item{stump.model}{object of class \code{prunedadaStump}}

\item{newdata}{Data to be fit}
}
\value{
A vector of length \code{nrow(newdata)} expressing the probability of ocurrence of the event modelled.
}
\description{
Predict from a \code{prunedadaStump} object
}
\examples{


#Load Iris
data(iris)

#Create Variable is Iris as numerical
iris$isSetosa <- as.numeric(iris$Species == "setosa")

#Split sample in 70 train - 30 test
train.ind <- sample(nrow(iris), nrow(iris) * 0.7)

#Train model. For obvious reasons, Species variable is not included in the fit
fit <- adaStump(isSetosa ~ Sepal.Length + Sepal.Width + Petal.Length + Petal.Width, iris[train.ind,], 
                type = "discrete", iter = 10, nu = 0.05, bag.frac = 0.6)
                
#Prune Tree and predict

fit_pruned <- pruneTree(fit)
predict(fit_pruned,iris[-train.ind,])

}

