% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruneTree.R
\name{pruneTree}
\alias{pruneTree}
\title{Merge stumps with repeated conditions}
\usage{
pruneTree(stump.model)
}
\arguments{
\item{stump.model}{Object of class "adaStump" to be reduced.}
}
\value{
An item of class \emph{prunedadaStump} containing the following:

\item{model }{a data.frame describing the pruned stumps} 
\item{type }{Type of ada execution performed.}
}
\description{
In order to reduce execution times, stumps with the same condition are collapsed. 
If a pruned model is used, n.iter can no longer be specified and all the stumps are used.
}
\examples{
#Load Iris
data(iris)

#Create Variable is Iris as numerical
iris$isSetosa <- as.numeric(iris$Species == "setosa")

#Split sample in 70 train - 30 test
train.ind <- sample(nrow(iris), nrow(iris) * 0.7)

#Train model. For obvious reasons, Species variable is not included in the fit
fit <- adaStump(isSetosa ~ Sepal.Length + Sepal.Width + Petal.Length + Petal.Width, iris[train.ind,], 
                type = "discrete", iter = 10, nu = 0.05, bag.frac = 0.6)
                
#Prune Tree and predict

fit_pruned <- pruneTree(fit)
predict(fit_pruned,iris[-train.ind,])
}

