\name{adaStump-package}
\alias{adaStump-package}
\alias{adaStump}
\docType{package}
\title{
\packageTitle{adaStump}
}
\description{
\packageDescription{adaStump}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{adaStump}
\packageIndices{adaStump}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{adaStump}

Maintainer: \packageMaintainer{adaStump}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%~~ simple examples of the most important functions ~~
}
