\name{adaStump-package}
\alias{adaStump-package}
\alias{adaStump}
\docType{package}
\title{
\packageTitle{adaStump}
}
\description{
\packageDescription{adaStump}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{adaStump}
\packageIndices{adaStump}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{adaStump}

Maintainer: \packageMaintainer{adaStump}
}